Fix Thread.start() that can hang indefinitely if the new thread fails (MemoryError) during its initialization
When we start a Thread (`Thread.start()`), it is possible that the current Thread will wait forever the signal of the Child Thread. That's possible if this new Thread doesn't have enough memory (or for another situation) and cannot signal that it starts (and that it finished in Error).

That's happen because `Thread.start()` waits `self._started.wait()` without timeout and without checking that the new threat is alive somehow.

Use the old timeout of Thread.start() used before 9e7f1d2e965400edcb2c0cb7fee625ef2b595eb5. When the timeout happens, look `_os_thread_handle` if the thread is dead somehow. If this is the case we can ensure that the thread has been killed without signal that it started.

IMO, this solution is not super clean (not proud of it) because it implies to have the old arbitrary timeout. But I don't see another easy solution right now.

Also change `Thread._delete()` to use pop instead, because nothing garantee that the Thread is in `_active[get_ident()]` exists. Indeed, `_bootstrap_inner` might crash before `_active[self._ident] = self`. Also use `_ident` because we know that `set_ident` has been already execute.